refactored final lambdas into separate files and removed db handle hard dependancies
