+++ chalice/chalicelib/demos.py
@@ -1,237 +1,19 @@
+"""
+TEST ROUTES START HERE - RUN AWS API ON DEMAND
+and demo routes with static data
+
+"""
+
 from datetime import datetime
-import json
-from chalicelib.database_handle import DatabaseHandle
 
+from chalice import Response
 
-def execute_test_ports_ingress_ssh(app, load_route_services):
+from app import app, load_route_services
+from chalicelib import models
+from chalicelib.template_handler import TemplateHandler
 
-    try:
-        load_route_services()
 
-        dbh = DatabaseHandle(app)
-
-        db = dbh.get_handle()
-        db.connect()
-
-        Criterion = dbh.get_model("Criterion")
-        Status = dbh.get_model('Status')
-
-        Client = app.utilities.get_class_by_name(
-            "chalicelib.criteria.aws_ec2_security_group_ingress_ssh.AwsEc2SecurityGroupIngressSsh"
-        )
-        ec2 = Client(app)
-
-        session = ec2.get_session(account='103495720024', role='csw-dan_CstSecurityInspectorRole')
-
-        region = 'eu-west-1'
-
-        groups = ec2.get_data(session, **{"region": region})
-
-        criterion = Criterion.get_by_id(1)
-
-        for group in groups:
-            compliance = ec2.evaluate({}, group, [])
-
-            app.log.debug(app.utilities.to_json(compliance))
-
-            group['resource_compliance'] = compliance
-
-            status = Status.get_by_id(compliance['status_id'])
-
-            group['status'] = status
-
-            group['resource_name'] = group['GroupName']
-            group['resource_id'] = group['GroupId']
-            group['region'] = region
-
-        summary = ec2.summarize(groups)
-
-        template_data = {
-            "criterion": criterion.serialize(),
-            "compliance_summary": summary,
-            "compliance_results": groups,
-            "tested": True
-        }
-
-        response = app.templates.render_authorized_route_template(
-            '/test/ports_ingress_ssh',
-            app.current_request,
-            template_data
-        )
-    except Exception as err:
-        response = {
-            "body": str(err)
-        }
-
-    return response
-
-
-def execute_test_ports_ingress_open(app, load_route_services):
-
-    try:
-        load_route_services()
-
-        Client = app.utilities.get_class_by_name(
-            "chalicelib.criteria.aws_ec2_security_group_ingress_open.AwsEc2SecurityGroupIngressOpen"
-        )
-        ec2 = Client(app)
-
-        session = ec2.get_session(account='103495720024', role='csw-dan_CstSecurityInspectorRole')
-
-        params = {
-            "region": 'eu-west-1'
-        }
-
-        groups = ec2.get_data(session, **params)
-
-        for group in groups:
-
-            compliance = ec2.evaluate({}, group, [])
-
-            app.log.debug(app.utilities.to_json(compliance))
-
-            group['resource_compliance'] = compliance
-
-        summary = ec2.summarize(groups)
-
-        template_data = app.dummy_data["audits"][0]
-        template_data["criteria"][0]["compliance_results"] = groups
-        template_data["criteria"][0]["compliance_summary"] = summary
-        template_data["criteria"][0]["tested"] = True
-
-        response = app.templates.render_authorized_route_template(
-            '/audit/{id}',
-            app.current_request,
-            template_data
-        )
-    except Exception as err:
-        response = {
-            "body": str(err)
-        }
-
-    return response
-
-
-def execute_test_root_mfa(app, load_route_services):
-
-    try:
-        load_route_services()
-
-        Client = app.utilities.get_class_by_name(
-            "chalicelib.criteria.aws_support_root_mfa.AwsSupportRootMfa"
-        )
-        support = Client(app)
-
-        session = support.get_session(
-            account='103495720024',
-            role='csw-dan_CstSecurityInspectorRole'
-        )
-
-        data = support.get_data(session)
-
-        criterion = {
-            "id": 3,
-            "criterion_name": "MFA enabled on root user account",
-            "description": "Checks whether the root IAM user associated with the AWS account has Multi Factor Authentication enabled.",
-            "why_is_it_important": "Without MFA it is easier for someone to gain access to your account",
-            "how_do_i_fix_it": "If you have the root credentials for your account enable MFA - otherwise speak to RE"
-        }
-
-        for item in data:
-            compliance = support.evaluate({}, item, [])
-
-            app.log.debug(app.utilities.to_json(compliance))
-
-            item['resource_compliance'] = compliance
-
-            status = {}
-
-            item['status'] = status
-
-            item.update(support.translate(item))
-
-        summary = support.summarize(data)
-
-        template_data = {
-            "criterion": criterion,
-            "compliance_summary": summary,
-            "compliance_results": data,
-            "tested": True
-        }
-
-        response = app.templates.render_authorized_route_template(
-            '/test/ports_ingress_ssh',
-            app.current_request,
-            template_data
-        )
-    except Exception as err:
-        response = {
-            "body": str(err)
-        }
-
-    return response
-
-
-def execute_test_iam_validate_inspector_policy(app, load_route_services):
-
-    try:
-        load_route_services()
-
-        Client = app.utilities.get_class_by_name(
-            "chalicelib.criteria.aws_iam_validate_inspector_policy.AwsIamValidateInspectorPolicy"
-        )
-        iam = Client(app)
-
-        session = iam.get_session(
-            account='103495720024',
-            role='csw-dan_CstSecurityInspectorRole'
-        )
-
-        data = iam.get_data(session)
-
-        criterion = {
-            "id": 4,
-            "criterion_name": "IAM inspector policy is up-to-date",
-            "description": "Checks whether the Cloud Security Watch role matches the current definition.",
-            "why_is_it_important": "If the role policy doesn't grant the right permissions checks will fail to be processed.",
-            "how_do_i_fix_it": "Update the module and re-run the terraform apply to re-deploy the role and policy."
-        }
-
-        for item in data:
-            compliance = iam.evaluate({}, item, [])
-
-            item['resource_compliance'] = compliance
-
-            status = {}
-
-            item['status'] = status
-
-            item.update(iam.translate(item))
-
-        summary = iam.summarize(data)
-
-        template_data = {
-            "criterion": criterion,
-            "compliance_summary": summary,
-            "compliance_results": data,
-            "tested": True
-        }
-
-        response = app.templates.render_authorized_route_template(
-            '/test/validate_iam_policy',
-            app.current_request,
-            template_data
-        )
-
-    except Exception as err:
-        response = {
-            "body": str(err)
-        }
-
-    return response
-
-
-audit = {
+AUDIT = {
     "name": "[User Name]",
     "audits": [
         {
@@ -311,3 +93,210 @@ audit = {
         }
     ]
 }
+
+
+@app.route('/demo')
+def demo_index():
+    app.dummy_data = AUDIT
+    load_route_services()
+    response = app.templates.render_authorized_route_template(
+        '/',
+        app.current_request,
+        {
+            "name": "[User Name]"
+        }
+    )
+    return Response(**response)
+
+
+@app.route('/demo/audit')
+def demo_audit_list():
+    app.dummy_data = AUDIT
+    load_route_services()
+    response = app.templates.render_authorized_route_template(
+        '/audit',
+        app.current_request,
+        app.dummy_data
+    )
+    return Response(**response)
+
+
+@app.route('/demo/audit/{id}')
+def demo_audit_report(id):
+    app.dummy_data = AUDIT
+    load_route_services()
+    app.templates = TemplateHandler(app)
+    response = app.templates.render_authorized_route_template(
+        '/audit/{id}',
+        app.current_request,
+        app.dummy_data["audits"][0]
+    )
+    return Response(**response)
+
+
+# RUN THE FOLLOWING AWS API ON DEMAND
+
+
+@app.route('/test/ports_ingress_ssh')
+def test_ports_ingress_ssh():
+    app.dummy_data = AUDIT
+    load_route_services()
+    try:
+        client = app.utilities.get_class_by_name(
+            "chalicelib.criteria.aws_ec2_security_group_ingress_ssh.AwsEc2SecurityGroupIngressSsh"
+        )
+        ec2 = client(app)
+        session = ec2.get_session(account='103495720024', role='csw-dan_CstSecurityInspectorRole')
+        region = 'eu-west-1'
+        groups = ec2.get_data(session, **{"region": region})
+        for group in groups:
+            compliance = ec2.evaluate({}, group, [])
+            app.log.debug(app.utilities.to_json(compliance))
+            group['resource_compliance'] = compliance
+            group['status'] = models.Status.get_by_id(compliance['status_id'])
+            group['resource_name'] = group['GroupName']
+            group['resource_id'] = group['GroupId']
+            group['region'] = region
+        summary = ec2.summarize(groups)
+        template_data = {
+            "criterion": models.Criterion.get_by_id(1).serialize(),
+            "compliance_summary": summary,
+            "compliance_results": groups,
+            "tested": True
+        }
+        response = app.templates.render_authorized_route_template(
+            '/test/ports_ingress_ssh',
+            app.current_request,
+            template_data
+        )
+    except Exception as err:
+        response = {
+            "body": str(err)
+        }
+    return Response(**response)
+
+
+@app.route('/test/ports_ingress_open')
+def test_ports_ingress_open():
+    app.dummy_data = AUDIT
+    load_route_services()
+    try:
+        client = app.utilities.get_class_by_name(
+            "chalicelib.criteria.aws_ec2_security_group_ingress_open.AwsEc2SecurityGroupIngressOpen"
+        )
+        ec2 = client(app)
+        session = ec2.get_session(account='103495720024', role='csw-dan_CstSecurityInspectorRole')
+        params = {
+            "region": 'eu-west-1'
+        }
+        groups = ec2.get_data(session, **params)
+        for group in groups:
+            compliance = ec2.evaluate({}, group, [])
+            app.log.debug(app.utilities.to_json(compliance))
+            group['resource_compliance'] = compliance
+        summary = ec2.summarize(groups)
+        template_data = app.dummy_data["audits"][0]
+        template_data["criteria"][0]["compliance_results"] = groups
+        template_data["criteria"][0]["compliance_summary"] = summary
+        template_data["criteria"][0]["tested"] = True
+        response = app.templates.render_authorized_route_template(
+            '/audit/{id}',
+            app.current_request,
+            template_data
+        )
+    except Exception as err:
+        response = {
+            "body": str(err)
+        }
+    return Response(**response)
+
+
+@app.route('/test/root_mfa')
+def test_root_mfa():
+    load_route_services()
+    try:
+        client = app.utilities.get_class_by_name(
+            "chalicelib.criteria.aws_support_root_mfa.AwsSupportRootMfa"
+        )
+        support = client(app)
+        session = support.get_session(
+            account='103495720024',
+            role='csw-dan_CstSecurityInspectorRole'
+        )
+        data = support.get_data(session)
+        criterion = {
+            "id": 3,
+            "criterion_name": "MFA enabled on root user account",
+            "description": "Checks whether the root IAM user associated with the AWS account has Multi Factor Authentication enabled.",
+            "why_is_it_important": "Without MFA it is easier for someone to gain access to your account",
+            "how_do_i_fix_it": "If you have the root credentials for your account enable MFA - otherwise speak to RE"
+        }
+        for item in data:
+            compliance = support.evaluate({}, item, [])
+            app.log.debug(app.utilities.to_json(compliance))
+            item['resource_compliance'] = compliance
+            status = {}
+            item['status'] = status
+            item.update(support.translate(item))
+        summary = support.summarize(data)
+        template_data = {
+            "criterion": criterion,
+            "compliance_summary": summary,
+            "compliance_results": data,
+            "tested": True
+        }
+        response = app.templates.render_authorized_route_template(
+            '/test/ports_ingress_ssh',
+            app.current_request,
+            template_data
+        )
+    except Exception as err:
+        response = {
+            "body": str(err)
+        }
+    return Response(**response)
+
+
+@app.route('/test/validate_iam_policy', cors=True)
+def validate_iam_policy():
+    load_route_services()
+    try:
+        client = app.utilities.get_class_by_name(
+            "chalicelib.criteria.aws_iam_validate_inspector_policy.AwsIamValidateInspectorPolicy"
+        )
+        iam = client(app)
+        session = iam.get_session(
+            account='103495720024',
+            role='csw-dan_CstSecurityInspectorRole'
+        )
+        data = iam.get_data(session)
+        criterion = {
+            "id": 4,
+            "criterion_name": "IAM inspector policy is up-to-date",
+            "description": "Checks whether the Cloud Security Watch role matches the current definition.",
+            "why_is_it_important": "If the role policy doesn't grant the right permissions checks will fail to be processed.",
+            "how_do_i_fix_it": "Update the module and re-run the terraform apply to re-deploy the role and policy."
+        }
+        for item in data:
+            compliance = iam.evaluate({}, item, [])
+            item['resource_compliance'] = compliance
+            status = {}
+            item['status'] = status
+            item.update(iam.translate(item))
+        summary = iam.summarize(data)
+        template_data = {
+            "criterion": criterion,
+            "compliance_summary": summary,
+            "compliance_results": data,
+            "tested": True
+        }
+        response = app.templates.render_authorized_route_template(
+            '/test/validate_iam_policy',
+            app.current_request,
+            template_data
+        )
+    except Exception as err:
+        response = {
+            "body": str(err)
+        }
+    return Response(**response)
